feat: better esc for fzf
Using C-c (and ESC if mapped to it) on windows quite often results in

^CTerminate batch job (Y/N)?

message and then error and everything kindof works after that...
But I don't like it.

C-g on the other hand make it work flawlessly.

PS, why map ESC as it should anyway close the terminal?
Probably I should remove regular tnoremap for esc.